Move origin whitelisting logic to own file
