+++ src/wrapNavigation.js
@@ -3,9 +3,8 @@
  *
  * Process: Main
  */
-
 const {app, shell, BrowserWindow} = require('electron');
-const {URL} = require('url');
+const {isUrlWhitelisted} = require('./isOriginWhitelisted');
 const {NAVIGATION_REQUESTED} = require('./channelNames');
 
 /**
@@ -32,7 +31,7 @@ function wrapNavigation() {
     // walled garden of whitelisted origins.
     // @see https://electron.atom.io/docs/api/web-contents/#event-will-navigate
     webContents.on('will-navigate', (event, url) => {
-      if (!isDestinationWhitelistedForNavigation(url)) {
+      if (!isUrlWhitelisted(url)) {
         event.preventDefault();
         shell.openExternal(url);
       }
@@ -52,7 +51,7 @@ function wrapNavigation() {
       // @see https://github.com/electron/electron/issues/1963
       event.preventDefault();
 
-      if (isDestinationWhitelistedForNavigation(url)) {
+      if (isUrlWhitelisted(url)) {
         const focusedWindow = BrowserWindow.getFocusedWindow();
         if (focusedWindow) {
           focusedWindow.webContents.send(NAVIGATION_REQUESTED, url);
@@ -65,40 +64,5 @@ function wrapNavigation() {
     });
   });
 }
-
-/**
- * @param {string} url
- * @return {boolean} TRUE if the target page is allowed to load within the
- *   Maker Toolkit Browser.
- */
-function isDestinationWhitelistedForNavigation(url) {
-  // Get origin from full URL
-  const origin = new URL(url).origin;
-
-  if (navigationBlacklist.some(blacklisted => blacklisted.test(origin))) {
-    return false;
-  }
-
-  // Match origins:
-  // https://studio.code.org                          Production
-  // https://test-studio.code.org                     Test
-  // https://dashboard-adhoc-my-branch.cdn-code.org   Ad-Hoc servers
-  // http://localhost-studio.code.org:3000            Local development
-  return /^https?:\/\/(?:[\w\d-]+\.)?(?:cdn-)?code\.org(?::\d+)?$/i.test(origin);
-}
-
-/**
- * Navigation to urls matching any of the given origins will open the page in
- * the system default browser, instead of within Maker Toolkit Browser, even
- * if the origin falls within our general Code.org origin whitelist.
- * @type {Array.<RegExp>}
- */
-const navigationBlacklist = [
-  /curriculum\.code\.org$/i,
-  /documentation\.code\.org$/i,
-  /forum\.code\.org$/i,
-  /support\.code\.org$/i,
-  /wiki\.code\.org$/i,
-];
 
 module.exports = wrapNavigation;
